+++ scandir/src/count/def.rs
@@ -1,5 +1,5 @@
 #[cfg(feature = "bincode")]
-use bincode::error::EncodeError;
+use bincode_next::error::EncodeError;
 #[cfg(feature = "speedy")]
 use speedy::{Readable, Writable};
 
@@ -58,7 +58,7 @@ impl Statistics {
 
     #[cfg(feature = "bincode")]
     pub fn to_vec(&self) -> Result<Vec<u8>, EncodeError> {
-        bincode::serde::encode_to_vec(self, bincode::config::legacy())
+        bincode_next::serde::encode_to_vec(self, bincode_next::config::legacy())
     }
 }