+++ scandir/src/scandir/def.rs
@@ -1,5 +1,5 @@
 #[cfg(feature = "bincode")]
-use bincode::error::EncodeError;
+use bincode_next::error::EncodeError;
 #[cfg(feature = "speedy")]
 use speedy::{Readable, Writable};
 
@@ -116,7 +116,7 @@ impl ScandirResult {
 
     #[cfg(feature = "bincode")]
     pub fn to_bincode(&self) -> Result<Vec<u8>, EncodeError> {
-        bincode::serde::encode_to_vec(self, bincode::config::legacy())
+        bincode_next::serde::encode_to_vec(self, bincode_next::config::legacy())
     }
 
     #[cfg(feature = "json")]
@@ -171,7 +171,7 @@ impl ScandirResults {
 
     #[cfg(feature = "bincode")]
     pub fn to_bincode(&self) -> Result<Vec<u8>, EncodeError> {
-        bincode::serde::encode_to_vec(self, bincode::config::legacy())
+        bincode_next::serde::encode_to_vec(self, bincode_next::config::legacy())
     }
 
     #[cfg(feature = "json")]