+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.10] - 2026-07-18
+
+### Fixed
+
+- Fix panic on invalid file names.
+
+### Changed
+
+- Switch from bincode to bincode-next.
+
 ## [2.9.9] - 2026-05-26
 
 ### Fixed